Uses the relations klass to define the root.
Currently it tries to infer the root of a collection from it's first element. However, if the collection is empty, the first element will be nil and the root will be nil.

Since it'll be given ActiveRecord::Relations most of the time, it now tries to infer the root from it's klass method first. Only if the collection has no klass method, it'll use the first element.