Adds scope
Instead of the helper module, a controller module gets included into controllers. It adds a scope as option from to controller to the serialization. By default the current_user will be the scope, but this can be custimized in the controller.

I've also changed all the keyword arguments to option hashes, since it's been getting too many variables.